requirements.txt
crewai
gradio
requests
beautifulsoup4
openai
python-dotenv
lxml
langchain-openai
